--- code/frontend/src/app/features/settings/orphaned-files-cleaner/orphaned-files-cleaner.component.html
@@ -1,130 +1,0 @@
-<app-page-header
-  title="Orphaned Files Cleaner"
-  subtitle="Move files not associated with any active torrent to a dedicated folder — currently supported with qBittorrent only"
-/>
-
-@if (loadError()) {
-  <app-empty-state
-    icon="tablerPlugConnectedX"
-    heading="Could not connect to server"
-    description="Failed to load settings. Please check that the server is running and try again."
-  >
-    <app-button variant="primary" size="sm" (clicked)="retry()">
-      Retry
-    </app-button>
-  </app-empty-state>
-} @else if (loader.showSpinner()) {
-  <app-loading-state message="Loading settings..." />
-} @else if (!loader.loading()) {
-<div class="settings-form">
-
-  <app-card header="Filters &amp; Safety">
-    <div class="form-stack">
-      <app-chip-input
-        label="Exclude Patterns"
-        placeholder="Add glob pattern (e.g. *.nfo)..."
-        hint="File or folder names matching these patterns are never considered orphaned (e.g. *.nfo, .DS_Store)"
-        [(items)]="excludePatterns"
-      />
-      <app-number-input
-        label="Min File Age"
-        suffix="min"
-        [min]="0"
-        hint="Skip files or folders modified less than this many minutes ago. Protects active downloads."
-        [(value)]="minFileAgeMinutes"
-      />
-      <app-number-input
-        label="Purge Orphaned After"
-        suffix="days"
-        [min]="1"
-        hint="Permanently delete entries from the Orphaned Directory after this many days. Leave empty to keep them indefinitely."
-        [(value)]="emptyAfterXDays"
-      />
-    </div>
-  </app-card>
-
-  <div class="form-actions">
-    <app-button
-      variant="primary"
-      [glowing]="dirty()"
-      [loading]="saving()"
-      [disabled]="saving() || saved() || !dirty()"
-      (clicked)="save()"
-    >
-      {{ saved() ? 'Saved!' : 'Save Settings' }}
-    </app-button>
-  </div>
-
-  @if (clientOptions().length > 0) {
-    <app-card [header]="selectedClient()?.downloadClientName ?? 'Download Client'">
-      <div class="form-stack">
-        <app-select
-          label="Download Client"
-          [options]="clientOptions()"
-          [value]="selectedClientId()"
-          (valueChange)="onClientChange($event)"
-          hint="Select which download client to configure"
-        />
-
-        @if (isSelectedClientDisabled()) {
-          <div class="disabled-client-notice">
-            This download client is disabled. Enable it in Download Client settings before configuring orphaned files.
-          </div>
-        }
-
-        @if (selectedClient(); as client) {
-          @if (!isSelectedClientDisabled()) {
-            <div class="form-divider"></div>
-
-            <app-toggle
-              label="Enabled"
-              [checked]="client.clientConfig?.enabled ?? false"
-              (checkedChange)="updateClientField('enabled', $event)"
-              hint="Enable orphaned files scanning for this download client"
-            />
-
-            @if (client.clientConfig?.enabled) {
-              <div class="form-divider"></div>
-
-              <app-chip-input
-                label="Scan Directories"
-                placeholder="Add directory path..."
-                hint="Absolute paths to scan for orphaned files. Each top-level entry is checked against your active torrents."
-                [items]="client.clientConfig?.scanDirectories ?? []"
-                (itemsChange)="updateClientField('scanDirectories', $event)"
-              />
-              <app-input
-                label="Orphaned Directory"
-                placeholder="/mnt/data/orphaned"
-                hint="Where orphaned files are moved. Leave empty to log only without moving."
-                [value]="client.clientConfig?.orphanedDirectory ?? ''"
-                (valueChange)="updateClientField('orphanedDirectory', $event || undefined)"
-              />
-
-            }
-
-            <div class="form-actions">
-              <app-button
-                variant="primary"
-                [glowing]="clientDirty()"
-                [loading]="clientSaving()"
-                [disabled]="clientSaving() || clientSaved() || !clientDirty()"
-                (clicked)="saveClientConfig()"
-              >
-                {{ clientSaved() ? 'Saved!' : 'Save Client Settings' }}
-              </app-button>
-            </div>
-          }
-        }
-      </div>
-    </app-card>
-  } @else {
-    <app-empty-state
-      icon="tablerDownload"
-      heading="No download clients"
-      description="Add a download client to configure per-client orphaned files settings"
-    />
-  }
-
-</div>
-}
