--- code/frontend/src/app/features/settings/orphaned-files-cleaner/orphaned-files-cleaner.component.scss
@@ -1,21 +1,0 @@
-@use 'settings-layout' as *;
-
-:host { @include settings-page; }
-
-.settings-form { @include settings-form; }
-.form-stack { @include form-stack; }
-.form-row { @include form-row; }
-.form-actions { @include form-actions; }
-.form-divider { @include form-divider; }
-
-.disabled-client-notice {
-  display: flex;
-  align-items: center;
-  gap: var(--space-2);
-  padding: var(--space-3) var(--space-4);
-  border-radius: var(--radius-md);
-  background: var(--color-warning-dim);
-  border: 1px solid var(--color-warning);
-  color: var(--color-warning);
-  font-size: var(--font-size-sm);
-}
